+++ src/koto/src/lib.rs
@@ -30,10 +30,7 @@ pub use {koto_bytecode as bytecode, koto_parser as parser, koto_runtime as runti
 
 use {
     koto_bytecode::{Chunk, LoaderError},
-    koto_runtime::{
-        DefaultStderr, DefaultStdout, KotoStderr, KotoStdout, Loader, MetaKey, RuntimeError, Value,
-        ValueMap, Vm, VmSettings,
-    },
+    koto_runtime::{KotoFile, Loader, MetaKey, RuntimeError, Value, ValueMap, Vm, VmSettings},
     std::{error::Error, fmt, path::PathBuf, sync::Arc},
 };
 
@@ -79,17 +76,20 @@ pub type KotoResult = Result<Value, KotoError>;
 pub struct KotoSettings {
     pub run_tests: bool,
     pub repl_mode: bool,
-    pub stdout: Arc<dyn KotoStdout>,
-    pub stderr: Arc<dyn KotoStderr>,
+    pub stdin: Arc<dyn KotoFile>,
+    pub stdout: Arc<dyn KotoFile>,
+    pub stderr: Arc<dyn KotoFile>,
 }
 
 impl Default for KotoSettings {
     fn default() -> Self {
+        let default_vm_settings = VmSettings::default();
         Self {
             run_tests: true,
             repl_mode: false,
-            stdout: Arc::new(DefaultStdout::default()),
-            stderr: Arc::new(DefaultStderr::default()),
+            stdin: default_vm_settings.stdin,
+            stdout: default_vm_settings.stdout,
+            stderr: default_vm_settings.stderr,
         }
     }
 }
@@ -120,6 +120,7 @@ impl Koto {
         Self {
             settings: settings.clone(),
             runtime: Vm::with_settings(VmSettings {
+                stdin: settings.stdin,
                 stdout: settings.stdout,
                 stderr: settings.stderr,
             }),